refactoring tv bloc
