+++ lib/ui/widgets/main_screen/main_screen_widget.dart
@@ -3,11 +3,11 @@ import 'package:comics_db_app/domain/blocs/movie/movie_list_state.dart';
 import 'package:comics_db_app/domain/blocs/movie/movie_popular_list_bloc.dart';
 import 'package:comics_db_app/domain/blocs/movie/now_playing_movie_list_bloc.dart';
 import 'package:comics_db_app/domain/blocs/movie/top_rated_movie_list_bloc.dart';
-import 'package:comics_db_app/domain/blocs/tv_airing_today_list_bloc.dart';
-import 'package:comics_db_app/domain/blocs/tv_on_the_air_bloc.dart';
-import 'package:comics_db_app/domain/blocs/tv_popular_list_bloc.dart';
-import 'package:comics_db_app/domain/blocs/tv_list_state.dart';
-import 'package:comics_db_app/domain/blocs/tv_top_rated_list_bloc.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_airing_today_list_bloc.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_on_the_air_bloc.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_popular_list_bloc.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_list_state.dart';
+import 'package:comics_db_app/domain/blocs/tv/tv_top_rated_list_bloc.dart';
 import 'package:comics_db_app/domain/blocs/movie/upcoming_movie_list_bloc.dart';
 import 'package:comics_db_app/domain/factories/screen_factory.dart';
 import 'package:comics_db_app/ui/widgets/movie_list/movie_list_cubit.dart';
